fix: restructure cli
