+++ __tests__/test.sh
@@ -8,12 +8,7 @@ chmod +x pkg/index.js
 ./pkg/index.js clean
 ./pkg/index.js build
 
-# Optimize
-chmod +x ./__tests__/optimize.sh
 ./__tests__/optimize.sh
-
-# Accessibility
-chmod +x ./__tests__/accessibility.sh
 ./__tests__/accessibility.sh
 
 # Generate components
@@ -26,7 +21,6 @@ else
 fi
 
 # Generate index files
-chmod +x ./__tests__/createIndexFiles.sh
 ./__tests__/createIndexFiles.sh
 
 # Clean directory
