+++ src/cli/__tests__/index.test.ts
