+++ src/cli/clean.ts
