+++ src/cli/createIndexFiles.ts
