+++ src/cli/helpers/nameHelper.ts
