+++ src/cli/optimize.ts
