+++ src/cli/svgoPlugins/pluginAddTitle.ts
