+++ src/cli/unknown.ts
@@ -1,4 +1,5 @@
 export const unknown = (availableCommands) => {
+  console.log({availableCommands})
   console.info(`Available commands: ${availableCommands.join(", ")}`)
   process.exit(1)
 }
